style: code organization
